Struct light_arena::Allocator

pub struct Allocator<'a> { /* fields omitted */ }

The allocator provides exclusive access to the memory arena, allowing for allocation of objects in the arena.

Objects allocated by an allocated cannot outlive it, upon destruction of the allocator the memory space it requested will be made available again. Drops of allocated objects are not called, only types which are Sized + Copy can be safely stored.

Methods

impl<'a> Allocator<'a>

pub fn alloc_slice<T: Sized + Copy>(&self, len: usize) -> &mut [T]

Get a dynamically sized slice of data from the allocator. The contents of the slice will be unintialized.