Struct light_arena::AllocatorPlacer

pub struct AllocatorPlacer<'a, T: 'a + Sized + Copy> { /* fields omitted */ }

Object representing a place to put a newly requested allocation.

Drop is never called so the placement new can only be run on Sized + Copy types. The lifetime of the place, and the subsequently placed T is tied to the lifetime of the Allocator which created the AllocatorPlacer.